feat: live prices, interactive charts, ranking, forming, narratives, alerts
- Live Upstox quotes: GET /api/v1/market-data/quotes + scan enrichment (current_price, change%)
- Interactive TradingView-style chart (lightweight-charts): candles + volume + S/R lines + B/R/C markers
- Quality scoring + ranking (top-N cards)
- Position sizing (account equity + risk %)
- Forming setup detection + watchlist
- Template narratives + invalidation copy
- Alert composer + Telegram delivery (ready)
- Scan history persistence + reload
- Product status banner (data freshness)
- Plug-and-play MARKET_DATA_SOURCE switching (demo/upstox)
- Volume BigInteger migration
- Updated PROJECT_PLAN.md to reflect all completed phases

+++ backend/app/infrastructure/database/models/candle.py
@@ -5,7 +5,7 @@ Uses Numeric for price fields and timezone-aware DateTime for timestamps.
 from datetime import datetime
 from decimal import Decimal
 from typing import Optional
-from sqlalchemy import Integer, ForeignKey, String, DateTime, Numeric, UniqueConstraint
+from sqlalchemy import Integer, BigInteger, ForeignKey, String, DateTime, Numeric, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 from ..base import Base
 
@@ -25,4 +25,4 @@ class CandleORM(Base):
     high: Mapped[Decimal] = mapped_column(Numeric(18, 8), nullable=False)
     low: Mapped[Decimal] = mapped_column(Numeric(18, 8), nullable=False)
     close: Mapped[Decimal] = mapped_column(Numeric(18, 8), nullable=False)
-    volume: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    volume: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
